fix updateProfile

+++ src/controllers/auth/updateImageProfile.ts
@@ -7,11 +7,13 @@ import deleteFile from '../services/deleteFile';
 
 const updateImageProfileUrl =async (req: Request, res: Response) => {
     try {
-        const oldpath = req.body.oldpath;
         const url = await generateSignedUrl("imageProfile",'my_profile');
-        const oldDirectory = oldpath.split("/")[3];
-        const oldFileName = oldpath.split("/")[4];
-        await deleteFile(oldDirectory,oldFileName);
+        if(req.body.oldpath){
+            const oldpath = req.body.oldpath;
+            const oldDirectory = oldpath.split("/")[3];
+            const oldFileName = oldpath.split("/")[4];
+            await deleteFile(oldDirectory,oldFileName);
+        }
         res.json({ url });
       } catch (error) {
         res.status(500).send('Error generating signed URL.');
